Make a dockerfile!
You can run this in a docker container locally by first building
  docker build -t learning-spring .
Then running the container. There's no hot reloading available afaik, so build every time before running
  docker run -p 8080:8080 learning-spring
Note: There's an issue at the moment. The server responds to the root url, but not to anything else, and there are no logs as to why. Even hitting the /guests endpoint and having a System.out.println doesn't show up in the terminal